collect weighting factor computations in one file mark tests in test_metric_fields.py which are duplicated in test_metrics_factory.py add TODO s for not registered fields
