+++ model/common/src/icon4py/model/common/metrics/compute_nudgecoeffs.py
@@ -30,6 +30,7 @@ def _compute_nudgecoeffs(
     )
 
 
+# TODO (@halungge) not registered in factory
 @program(grid_type=GridType.UNSTRUCTURED)
 def compute_nudgecoeffs(
     nudgecoeffs_e: fa.EdgeField[wpfloat],
