+++ model/common/tests/metric_tests/test_compute_coeff_gradekin.py
@@ -14,6 +14,7 @@ from icon4py.model.common.metrics.compute_coeff_gradekin import compute_coeff_gr
 from icon4py.model.testing import datatest_utils as dt_utils, helpers
 
 
+@pytest.mark.duplicate
 @pytest.mark.datatest
 @pytest.mark.parametrize("experiment", [dt_utils.REGIONAL_EXPERIMENT, dt_utils.GLOBAL_EXPERIMENT])
 def test_compute_coeff_gradekin(icon_grid, grid_savepoint, metrics_savepoint):
